backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9  # если используете PostgreSQL
django-cors-headers>=4.3